Compilation and somewhat running on macOS

+++ sources/etx/rt/integrators/vcm_gpu.cxx
@@ -3,6 +3,8 @@
 
 #include <etx/rt/shared/vcm_shared.hxx>
 
+#include <unordered_map>
+
 namespace etx {
 
 #if (0)
@@ -336,7 +338,7 @@ struct GPUVCMImpl {
     total_time = {};
 
     rt.gpu()->destroy_buffer(global_data);
-    global_data = rt.gpu()->create_buffer({align_up(sizeof(VCMIteration), 16llu) + align_up(sizeof(VCMGlobal), 16llu)});
+    global_data = rt.gpu()->create_buffer({align_up(sizeof(VCMIteration), size_t(16)) + align_up(sizeof(VCMGlobal), size_t(16))});
 
     rt.gpu()->destroy_buffer(state_buffers[0]);
     state_buffers[0] = rt.gpu()->create_buffer({sizeof(VCMPathState) * dimemsions.x * dimemsions.y});
